move ScalarType
